+++ Sources/Hyperspace/Service/Network/NetworkActivityIndicatable.swift
@@ -1,6 +1,6 @@
 //
 //  NetworkActivityIndicatable.swift
-//  Hyperspace-iOS
+//  Hyperspace
 //
 //  Created by William McGinty on 12/21/17.
 //  Copyright © 2017 Bottle Rocket Studios. All rights reserved.